BISA PINDAH2 PETA YAY

+++ ADT/point/point.h
@@ -8,8 +8,8 @@
 #include "../boolean/boolean.h"
 
 typedef struct {
-	float X; /* absis   */
-	float Y; /* ordinat */
+	int X; /* absis   */
+	int Y; /* ordinat */
 } POINT;
 
 /* *** Notasi Akses: Selektor POINT *** */
